Split up completions into separate files
